podcast-tracker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
feedparser==6.0.11
apscheduler==3.10.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
